refactor: rename actions module to action
