+++ src/ui/mod.rs
@@ -1,6 +1,6 @@
 //! Plugin, components, systems, and events related to common UI patterns.
 
-pub mod actions;
+pub mod action;
 pub mod button_radial;
 pub mod font_scale;
 pub mod icon_button;
